% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{esa.archives}
\alias{esa.archives}
\title{Downloads supplementary materials from Ecological Archives}
\usage{
esa.archives(esa, si.name, data.paper = TRUE, save.name = NULL,
  dir = NULL, cache = TRUE)
}
\arguments{
\item{esa}{ESA code for the article (just below the title in the
manuscript)}

\item{si.name}{*name* of the file to be downloaded}

\item{data.paper}{whether paper is an ESA data paper (default:
TRUE), or you are attempting to download the supplement to a
standard paper (e.g., one in Ecology)}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}

\item{cache}{if \code{TRUE} (default), the file won't be downloaded
again if it already exists (in the temporary directory \code{grabr}
creates, or your chosen \code{dir})}
}
\description{
Downloads supplementary materials from Ecological Archives
}
\examples{
fungi <- read.csv(esa.archives("E093-059", "myco_db.csv"))
mammals <- read.csv(esa.archives("E092-201", "MCDB_communities.csv", FALSE))
}
\author{
Will Pearse
}

