% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{figshare}
\alias{figshare}
\title{Downloads supplementary materials from FigShare}
\usage{
figshare(doi, si.no, save.name = NULL, dir = NULL, cache = TRUE)
}
\arguments{
\item{doi}{DOI of article}

\item{si.no}{which supplement (first, second, etc.) as printed in
the article, to download. Note that "Fig S1" is not valid; this
should be a number}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}

\item{cache}{if \code{TRUE} (default), the file won't be downloaded
again if it already exists (in the temporary directory \code{grabr}
creates, or your chosen \code{dir})}
}
\description{
Downloads supplementary materials from FigShare
}
\examples{
figshare("10.6084/m9.figshare.979288", 1)
}
\author{
Will Pearse
}

