% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{plos}
\alias{plos}
\title{Downloads supplementary materials from PLoS journals}
\usage{
plos(doi, si.no, save.name = NULL, dir = NULL, cache = TRUE)
}
\arguments{
\item{doi}{DOI of article}

\item{si.no}{which supplement (first, second, etc.) as printed in
the article, to download. Note that "Fig S1" is not valid; this
should be a number}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}

\item{cache}{if \code{TRUE} (default), the file won't be downloaded
again if it already exists (in the temporary directory \code{grabr}
creates, or your chosen \code{dir})}
}
\description{
Downloads supplementary materials from PLoS journals
}
\examples{
plos("10.1371/journal.pone.0127900", 1)
plos("10.1371/journal.pbio.1002177", 1)
plos("10.1371/journal.pmed.1001847", 1)
plos("10.1371/journal.pgen.1005291", 1)
plos("10.1371/journal.pcbi.1004327", 1)
plos("10.1371/journal.ppat.1005005", 1)
plos("10.1371/journal.pntd.0003824", 1)
}
\author{
Will Pearse
}

