% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{proceedings}
\alias{proceedings}
\title{Downloads supplementary materials from Proceedings journals}
\usage{
proceedings(doi, si.no, vol, issue, save.name = NULL, dir = NULL)
}
\arguments{
\item{doi}{DOI of article}

\item{si.no}{which supplement (first, second, etc.) as printed in
the article, to download. Note that "Fig S1" is not valid; this
should be a number}

\item{vol}{volume of article}

\item{issue}{issue of article}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}
}
\description{
Downloads supplementary materials from Proceedings journals
}
\examples{
proceedings("10.1098/rspb.2015.0338", 282, 1811, 1)
}
\author{
Will Pearse
}

