% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{science}
\alias{science}
\title{Downloads supplementary materials from Science}
\usage{
science(doi, si.name, save.name = NULL, dir = NULL)
}
\arguments{
\item{doi}{DOI of article}

\item{si.name}{*name* of the file to be downloaded}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}
}
\description{
Downloads supplementary materials from Science
}
\examples{
science("10.1126/science.1255768", "Appendix_BanksLeite_etal.txt")
}
\author{
Will Pearse
}

