\name{suppdata-package}
\alias{suppdata-package}
\alias{suppdata}
\docType{package}
\title{
\packageTitle{suppdata}
}
\description{
\packageDescription{suppdata}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{suppdata}
\packageIndices{suppdata}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{suppdata}

Maintainer: \packageMaintainer{suppdata}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
}
