% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{wiley}
\alias{wiley}
\title{Downloads supplementary materials from Wiley journals}
\usage{
wiley(doi, si.no, save.name = NULL, dir = NULL)
}
\arguments{
\item{doi}{DOI of article}

\item{si.no}{which supplement (first, second, etc.) as printed in
the article, to download. Note that "Fig S1" is not valid; this
should be a number}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}
}
\description{
Downloads supplementary materials from Wiley journals
}
\examples{
wiley("10.1111/ele.12437", 1)
}
\author{
Will Pearse
}

