% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xeno.R
\name{xeno.canto.download}
\alias{xeno.canto.download}
\title{Xeno-Canto data download wrapper}
\usage{
xeno.canto.download(ids, save.name = NULL, dir = NULL)
}
\arguments{
\item{ids}{vector of ids (*not* URLs), probably taken from
\code{\link{xeno.canto.meta}}}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}
}
\description{
Parameters and meanings are described in Xeno-Canto API
documentation (http://www.xeno-canto.org/article/153). This is a
very light wrapper! You can download the next page of data using
the 'page' parameter; I've yet to write a 'continue downloading'
wrapper, in part for fear of over-using the API.
}
\examples{
xeno.canto.download(xeno.canto.meta(setNames(c("bearded bellbird",
    "A"), c("species", "q")))$id[1])
xeno.canto.download("247117")
xeno.canto.download(247117:247120)
}
\author{
Will Pearse
}
