% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xeno.R
\name{xeno.canto.meta}
\alias{xeno.canto.meta}
\title{Xeno-Canto meta-data download wrapper}
\usage{
xeno.canto.meta(params, verbose = TRUE, types = c("id", "gen", "sp", "ssp",
  "en", "rec", "cnt", "loc", "lat", "lng", "type", "file", "lic", "url", "q"))
}
\arguments{
\item{params}{named vector of parameters to search Xeno-Canto
with. Names should be the parameter type, and values should be the
values for those searches.}

\item{verbose}{whether to print what page (of how many) is being
downloaded (default: TRUE)}

\item{types}{categories of meta-data to return about
calls. Defaults are intended to be everything everything; for space
reasons you might not want everything.}
}
\description{
Parameters and meanings are described in Xeno-Canto API
documentation (http://www.xeno-canto.org/article/153). This is a
very light wrapper! You can download the next page of data using
the 'page' parameter; I've yet to write a 'continue downloading'
wrapper, in part for fear of over-using the API.
}
\examples{
xeno.canto(setNames(c("bearded bellbird", "A"), c("species", "q")))
}
\author{
Will Pearse
}

