% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/journals.R
\name{esa.archives}
\alias{esa.archives}
\title{Downloads supplementary materials from Ecological Archives}
\usage{
esa.archives(esa, si.name, save.name = NULL, dir = NULL)
}
\arguments{
\item{esa}{ESA code for the article (just below the title in the
manuscript)}

\item{si.name}{*name* of the file to be downloaded}

\item{save.name}{a name for the file to download. If \code{NULL}
(default) this will be a combination of the DOI and SI number}

\item{dir}{directory to save file to. If \code{NULL} (default) this
will be a temporary directory created for your files}
}
\description{
Downloads supplementary materials from Ecological Archives
}
\examples{
esa.archives("E093-059", "myco_db.csv")
}
\author{
Will Pearse
}

