\name{grabr-package}
\alias{grabr-package}
\alias{grabr}
\docType{package}
\title{
\packageTitle{grabr}
}
\description{
\packageDescription{grabr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{grabr}
\packageIndices{grabr}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{grabr}

Maintainer: \packageMaintainer{grabr}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
